kim resigned, moving her page to /people/ex
"Hello all,

I am resigning from the Art Student Council as I do not see myself fit
for a position. Since Sebastian and I were the two students on the
Academic and Student Affairs Committee and we have both resigned, new
positions will have to be filled. The dates for the meetings, which are
held in LL101, are as follows:

           February 13, 12:45 p.m. – 2:45 p.m.
           May 15, 12:45 p.m. – 2:45 p.m.
           September 4, 12:45 p.m. – 2:45 p.m.
           November 20, 12:45 p.m. – 2:45 p.m.

I wish the Art Student Council luck in it's endeavors for transparency
with the faculty and the student body, as well as the effort to unify
the student body.

All the Best,
Kim"