shira retires, with highest regards and tallest respect
Hey ASC!

For those of you who do not already know, I am formally stepping out
for a semester leave from ASC. I do not know for sure whether or not I
will be back, with hopes that I will, but we will cross that bridge
when we get there. For now, I am finding meaning in my work more than I
have before at while at Cooper, and I want to focus on it and on other
out-of-school parts of my life that I need to give time to. Apologies
for being sluggish and emailing, I was talking to those of you in
person that I was able to, and the email helps so that official changes
can be made, and ASC can move on wit' it.

Senate is meeting on February 26th, and I am currently on the committee
and RSVPd, FYI. I can let them know I am not participating this
semester, or whatever is best.

I appreciate you all so much!! And I'm not stepping out with any hard
feelings - you're all wonderful and have put so much time and energy
into such a necessary part of school, and it's awesome to see the new
ways it is functioning.

PS - TAG! Jon's it!

With highest regards and tallest respect,
Shira Service
Former ASC Rep
RIP, BRB